threading::pool
Struct ThreadPool

pub struct ThreadPool { /* private fields */ }

A thread pool that manages a set of worker threads to execute tasks concurrently.

The ThreadPool allows you to submit closures to be executed in separate threads,
and provides a way to retrieve the results of those closures through Future objects.

Implementations

impl ThreadPool

pub fn new(thread_count: u16) -> Self

Creates a new ThreadPool with the specified number of worker threads.

pub fn run<F, O>(&self, task: F) -> ArcFuture<O>
where F: FnOnce() -> O + Send + UnwindSafe + 'static, O: Send + 'static,

Submits a closure to be executed by a worker thread and returns a Future for the result.

Parameters

task: The closure to be executed. It must be Send, 'static, and panic-safe.

Returns

An ArcFuture<O> that can be used to retrieve the result of the closure once it has completed.

pub fn flying_tasks_count(&self) -> u16

Returns the number of not yet exectued tasks in the pool.