threading
Struct Channel

pub struct Channel<R, W> { /* private fields */ }

Implementations

impl<R, W> Channel<R, W>

pub fn new_pair() -> (Channel<W, R>, Channel<R, W>)

Creates a new pair of inter-connected channels

impl<R: PartialEq, W: PartialEq> Channel<R, W>

pub fn wait_for(&self, waited_message: R) -> Result<(), RecvError>

Waits for a specific message of type R to be received.

Parameters

waited_message: The message to wait for.

Returns

Ok(()) if the waited message is received.

An error of type std::sync::mpsc::RecvError if the receiving fails.

pub fn wait_for_or_timeout( &self, waited_message: R, timeout: Duration, ) -> Result<(), RecvTimeoutError>

Waits for a specific message of type R to be received or times out after a specified duration.

Parameters

waited_message: The message to wait for.

timeout: The duration to wait before timing out.

Returns

Ok(()) if the waited message is received before the timeout.

An error of type std::sync::mpsc::RecvTimeoutError::Timeout if the timeout occurs.

An error of type std::sync::mpsc::RecvTimeoutError if the receiving fails.

pub fn send(&self, t: W) -> Result<(), SendError<W>>

Short hand for std::sync::mpsc::Sender::send

pub fn iter(&self) -> Iter<'_, R>

Short hand for std::sync::mpsc::Receiver::iter

pub fn try_iter(&self) -> TryIter<'_, R>

Short hand for std::sync::mpsc::Receiver::try_iter

pub fn recv(&self) -> Result<R, RecvError>

Short hand for std::sync::mpsc::Receiver::recv

pub fn try_recv(&self) -> Result<R, TryRecvError>

Short hand for std::sync::mpsc::Receiver::try_recv

pub fn recv_timeout(&self, timeout: Duration) -> Result<R, RecvTimeoutError>

Short hand for std::sync::mpsc::Receiver::recv_timeout